+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## blissify 0.2.2
+* Add a `list-db` subcommand to list what was analyzed.
+* Make blissify toggle random mode off when making playlists.
+
 ## blissify 0.2.1
 * Add a proper progressbar for the analysis.